Integ tests: export _get_compute_nodes_allocation to common
